Moving UUID generation code to new arrow.
