+++ CMake/kwiver-depends.cmake
@@ -14,3 +14,4 @@ include( kwiver-depends-Matlab )
 include( kwiver-depends-TinyXML )
 include( kwiver-depends-darknet )
 include( kwiver-depends-Burn-Out )
+include( kwiver-depends-uuid )
